django/test-requirements.txt
Django==1.4.3
factory-boy==1.2.0
